#11 Move ID class (typing → specs)
